move config components to configs folder
