+++ apps/portals-example/src/components/configs/ComponentsConfig/ComponentsConfig.tsx
